+++ openlibrary/tests/core/test_processors.py
@@ -3,6 +3,7 @@ import re
 import web
 
 from infogami.infobase import client, common
+from openlibrary.core import helpers
 from openlibrary.core.processors import readableurls as processors
 
 
@@ -145,8 +146,8 @@ def _make_thing(site, key, type_key, extra=None):
 
 def test_is_exclusion_author_excluded(monkeypatch):
     """An author whose OLID is in author_exclusions returns True."""
-    monkeypatch.setattr(processors, "_AUTHOR_EXCLUSIONS", None)
-    monkeypatch.setattr(processors.config, "get", lambda key, *a, **kw: ["OL99A"] if key == "author_exclusions" else None)
+    monkeypatch.setattr(helpers, "_AUTHOR_EXCLUSIONS", None)
+    monkeypatch.setattr(helpers.config, "get", lambda key, *a, **kw: ["OL99A"] if key == "author_exclusions" else None)
     site = MockSite()
     author = _make_thing(site, "/authors/OL99A", "/type/author", {"name": "Test Author"})
     assert processors.is_exclusion(author) is True
@@ -154,8 +155,8 @@ def test_is_exclusion_author_excluded(monkeypatch):
 
 def test_is_exclusion_author_not_excluded(monkeypatch):
     """An author not in author_exclusions returns False."""
-    monkeypatch.setattr(processors, "_AUTHOR_EXCLUSIONS", None)
-    monkeypatch.setattr(processors.config, "get", lambda key, *a, **kw: [] if key == "author_exclusions" else None)
+    monkeypatch.setattr(helpers, "_AUTHOR_EXCLUSIONS", None)
+    monkeypatch.setattr(helpers.config, "get", lambda key, *a, **kw: [] if key == "author_exclusions" else None)
     site = MockSite()
     author = _make_thing(site, "/authors/OL1A", "/type/author", {"name": "Safe Author"})
     assert processors.is_exclusion(author) is False
@@ -163,8 +164,8 @@ def test_is_exclusion_author_not_excluded(monkeypatch):
 
 def test_is_exclusion_empty_config_returns_false(monkeypatch):
     """When author_exclusions is absent from config, nothing is excluded."""
-    monkeypatch.setattr(processors, "_AUTHOR_EXCLUSIONS", None)
-    monkeypatch.setattr(processors.config, "get", lambda key, *a, **kw: None)
+    monkeypatch.setattr(helpers, "_AUTHOR_EXCLUSIONS", None)
+    monkeypatch.setattr(helpers.config, "get", lambda key, *a, **kw: None)
     site = MockSite()
     author = _make_thing(site, "/authors/OL99A", "/type/author", {"name": "Test Author"})
     assert processors.is_exclusion(author) is False
@@ -172,14 +173,14 @@ def test_is_exclusion_empty_config_returns_false(monkeypatch):
 
 def test_is_exclusion_none_returns_false(monkeypatch):
     """is_exclusion(None) returns False without raising."""
-    monkeypatch.setattr(processors, "_AUTHOR_EXCLUSIONS", None)
-    monkeypatch.setattr(processors.config, "get", lambda key, *a, **kw: [])
+    monkeypatch.setattr(helpers, "_AUTHOR_EXCLUSIONS", None)
+    monkeypatch.setattr(helpers.config, "get", lambda key, *a, **kw: [])
     assert processors.is_exclusion(None) is False
 
 
 def test_get_author_exclusions_caches(monkeypatch):
-    """_get_author_exclusions returns the same frozenset on repeated calls."""
-    monkeypatch.setattr(processors, "_AUTHOR_EXCLUSIONS", None)
+    """get_author_exclusions returns the same frozenset on repeated calls."""
+    monkeypatch.setattr(helpers, "_AUTHOR_EXCLUSIONS", None)
     call_count = {"n": 0}
 
     def fake_get(key, *a, **kw):
@@ -188,8 +189,8 @@ def test_get_author_exclusions_caches(monkeypatch):
             return ["OL1A", "OL2A"]
         return None
 
-    monkeypatch.setattr(processors.config, "get", fake_get)
-    r1 = processors._get_author_exclusions()
-    r2 = processors._get_author_exclusions()
+    monkeypatch.setattr(helpers.config, "get", fake_get)
+    r1 = helpers.get_author_exclusions()
+    r2 = helpers.get_author_exclusions()
     assert r1 is r2  # same object — cache hit
     assert call_count["n"] == 1  # config.get called only once
